requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
Jinja2==3.1.6
PyYAML==6.0.2
colorlog==6.9.0
luadata==1.0.5
slowapi==0.1.9
psutil==7.0.0
cryptography==44.0.2